feat: source-wise eval, run provenance, and measured wall-clock
The audit in reports/20260729.md found the efficiency and accuracy claims
resting on numbers that the tooling could not actually produce.

eval.py read one concatenated stream and truncated at --max-clips, so
"synthetic 500 clips" was VKITTI2 500 clips. It now builds a loader per
source, counts --max-clips per source, and separates the two aggregates
that were being conflated: MEAN(src) weights datasets equally, POOLED(px)
weights pixels. Active ratio turns out to span 14.5% (TUM) to 70.3%
(TartanAir), so the single "15.4% active" figure was one source's value.
Added the unscaled columns too (mAbsRel / scale / drift): median scaling
scores relative structure only, and a fixed camera needs metres.

infer.py's --spatial-cache was a store_true whose False default overrode
the checkpoint's trained spatial_cache=true, and --size 128 overrode a
256px run, so the documented command silently took the dense path at the
wrong resolution. Both flags are tri-state now and the resolution comes
from the sidecar config via the new checkpoint_config().

train.py gained --seed, and writes the *effective* config (merged CLI+TOML,
plus git commit, versions and the command) instead of copying the input
TOML — arm1/arm2 recorded teacher_weight = 0.5 for runs that overrode it
to 0. The same meta travels inside the checkpoint.

bench.py is new: latency/FPS against a forced active ratio, cache on/off,
fp16, multi-stream, peak VRAM and per-stream state. Measuring with it
exposed two deployment bugs, both fixed here: the detector emits fp32
masks whatever the model's dtype, and multiplying delta by one promoted
the whole scan back to fp32, which made fp16 inference impossible; and the
4-way scan's permutation tensor was first created inside the captured
region, so cudagraph replays clobbered the entry the cache still pointed
at.

+++ sokkanaem/ssm.py
@@ -46,7 +46,11 @@ class SelectiveSSM(nn.Module):
         x, z = self.in_proj(u).chunk(2, dim=-1)          # (B, L, d_inner)
         dt = F.softplus(self.dt_proj(x))                  # (B, L, d_inner)
         if mask is not None:
-            dt = dt * mask.unsqueeze(-1)                  # Δ-gating: the core trick
+            # .to(dt): the detector builds masks in fp32 regardless of the
+            # model's dtype, and multiplying by one promotes the whole scan
+            # back to fp32 — half the tensors downstream then mismatch and
+            # a half-precision model cannot run at all
+            dt = dt * mask.unsqueeze(-1).to(dt.dtype)     # Δ-gating: the core trick
         Bp, Cp = self.bc_proj(x).chunk(2, dim=-1)         # (B, L, d_state)
         return x, z, dt, Bp, Cp
